+++ packages/fes-builder-webpack/src/index.js
@@ -24,6 +24,7 @@ export default function () {
             require.resolve('./plugins/features/postcssLoader'),
             require.resolve('./plugins/features/nodeModulesTransform'),
             require.resolve('./plugins/features/vueLoader'),
+            require.resolve('./plugins/features/extraCSS'),
 
             // commands
             require.resolve('./plugins/commands/build'),
