+++ packages/fes-builder-webpack/src/plugins/common/buildDevUtils.js
@@ -5,8 +5,8 @@
 
 import { join, resolve } from 'path';
 import { existsSync, readFileSync } from 'fs';
-import { rimraf, chalk } from '@fesjs/utils';
 import zlib from 'zlib';
+import { rimraf, chalk } from '@fesjs/utils';
 import getConfig from './webpackConfig';
 
 export async function getBundleAndConfigs({ api }) {
