+++ packages/fes-builder-webpack/src/plugins/common/webpackConfig/getBabelOpts.js
