+++ packages/fes-builder-webpack/src/plugins/common/webpackConfig/minimizer.js
