+++ packages/fes-builder-webpack/src/plugins/common/webpackConfig/vue.js
