+++ packages/fes-builder-webpack/src/plugins/features/extraCSS.js
@@ -1,0 +1,14 @@
+export default (api) => {
+    api.describe({
+        key: 'extraCSS',
+        config: {
+            schema(joi) {
+                return joi.object({
+                    plugin: joi.object(),
+                    loader: joi.object(),
+                });
+            },
+            default: {},
+        },
+    });
+};
